Time: 122 ms (76.78%), Space: 62.9 MB (73.93%) - LeetHub

+++ 2530-maximal-score-after-applying-k-operations/2530-maximal-score-after-applying-k-operations.java
@@ -8,7 +8,7 @@ class Solution {
         while(i<k) {
             int n = pq.remove();
             sum+=n;
-            pq.add((int)Math.ceil(n/3.0));
+            pq.add((int)Math.ceil((double)n/3));
             i++;
         }
